+++ src/actions/clone.rs
@@ -1,7 +1,7 @@
 /*
  * Copyright 2023 EAS Barbosa
  *
- *     Licensed under the Apache License, Version 2.0(the "License");
+ *     Licensed under the Apache Licens0ersion 2.0(the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
@@ -13,3 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+use std::path::PathBuf;
+
+use git2::Repository;
+
+pub fn run(project: &crate::domain::project::Project, path: PathBuf) {
+    match Repository::clone(&project.url, path) {
+        Ok(repo) => repo,
+        Err(e) => panic!("failed to clone: {}", e),
+    };
+}